:bug: :white_check_mark: fix tests

See merge request lbb/lbb!649

+++ docker/v3.10/entrypoints/test.sh
@@ -1,36 +1,42 @@
 # /bin/bash
 
-# run custom sql scripts if any
-echo 'Running sql scripts if any'
-for i in `/bin/ls -1 /sql/*.sql`; do 
-    echo $i
-    mysql --user=$DB_USER \
-    --password=$DB_PASSWORD \
-    --host=$DB_HOST \
-    --port=$DB_PORT \
-    --database=$DB_NAME < $i
-done
+testReturn=0
 
-# create the index in elastic search
-poetry run create_index --full
+failed() {
+    testReturn=1
+    echo "FAILED $1"
+}
+
+set -ex
+
+# alembic
+poetry run alembic -c labonneboite/alembic.ini upgrade head
 
 # run the tests
 # -- lint
-poetry run flake8 --output-file flake8.txt || echo "FAILED flake"
-poetry run flake8_junit flake8.txt flake8.xml
-rm flake8.txt
+if ! poetry run flake8 --output-file flake8.txt; then
+    failed "flake"
+    poetry run flake8_junit flake8.txt flake8.xml && echo
+    rm flake8.txt
+fi
 
 # -- unit test & coverage
 # -- api
-poetry run pytest --junitxml=pytest-web-api.xml --cov --html=pytest-web-api.html
+if ! poetry run pytest --junitxml=pytest-web-api.xml --cov --html=pytest-web-api.html --maxfail=${MAXFAIL:-1000}; then
+    failed "pytest"
+fi
 poetry run coverage xml
 
 # -- build package
-poetry build
+if ! poetry build; then
+    failed "build"
+fi
 
 # prepare test results
 echo "Moving test results file..."
 mkdir -p testResults
-mv *.xml  ./testResults
-mv *.html  ./testResults
+mv *.xml ./testResults
+mv *.html ./testResults
 echo "Done"
+
+exit $testReturn
